Auto cmd format is now "Auto"+StartLocation+"To"+Target1+"To"+Target2...
New commands were added using this format and it makes more sense given
we are now coding more complex autos.  This commit changes all the autos
to match that format.  For example:

Start on the Left -> Go to Left switch -> Go to Right scale would be

AutoLeftToLeftSwitchToRightScale

More details can be appended to the end as necessary (e.g. to indicate
an approach from a side or back)